refactor: config로 이름 변경
