+++ big-trader-api/src/test/kotlin/com/flab/bigtrader/common/IntegrationTestConfig.java
@@ -1,0 +1,8 @@
+package com.flab.bigtrader.common;
+
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration
+public class IntegrationTestConfig {
+
+}
